Server robots worker
This worker will be used to serve robots.txt files accross our infrastructure.

The worker should be set on *.example.com/robots.txt (and optionally example.com/robots.txt). Each robot file is defined in the  directory, the file name being the domain the rule should be served on.

The rules will be defined in a latter PR, right now only the main site is set up.